+++ stdb/query/BUILD
@@ -38,6 +38,23 @@ cc_library(
     "query_processing/sliding_window.h",
     "query_processing/spacesaver.h",
     "query_processing/top.h",
+
+    "plan/aggregate_combiner.h",
+    "plan/aggregate.h",
+    "plan/aggregate_processing_step.h",
+    "plan/chain.h",
+    "plan/filter_processing_step.h",
+    "plan/group_aggregate_combiner.h",
+    "plan/group_aggregate_filter_processing_step.h",
+    "plan/group_aggregate_processing_step.h",
+    "plan/join.h",
+    "plan/materialization_step.h",
+    "plan/mergeby.h",
+    "plan/processing_prelude.h",
+    "plan/scan_events_processing_step.h",
+    "plan/scan_processing_step.h",
+    "plan/series_order_aggregate.h",
+    "plan/time_order_aggregate.h",
   ],
   alwayslink = 1,
   copts = [
