+++ stdb/query/queryplan.cc
@@ -3,886 +3,25 @@
  */
 #include "stdb/query/queryplan.h"
 
-#include "stdb/storage/nbtree.h"
-#include "stdb/storage/column_store.h"
-#include "stdb/storage/operators/operator.h"
-#include "stdb/storage/operators/scan.h"
-#include "stdb/storage/operators/merge.h"
-#include "stdb/storage/operators/aggregate.h"
-#include "stdb/storage/operators/join.h"
+#include "stdb/query/plan/aggregate_combiner.h"
+#include "stdb/query/plan/aggregate.h"
+#include "stdb/query/plan/aggregate_processing_step.h"
+#include "stdb/query/plan/chain.h"
+#include "stdb/query/plan/filter_processing_step.h"
+#include "stdb/query/plan/group_aggregate_combiner.h"
+#include "stdb/query/plan/group_aggregate_filter_processing_step.h"
+#include "stdb/query/plan/group_aggregate_processing_step.h"
+#include "stdb/query/plan/join.h"
+#include "stdb/query/plan/materialization_step.h"
+#include "stdb/query/plan/mergeby.h"
+#include "stdb/query/plan/processing_prelude.h"
+#include "stdb/query/plan/scan_events_processing_step.h"
+#include "stdb/query/plan/scan_processing_step.h"
+#include "stdb/query/plan/series_order_aggregate.h"
+#include "stdb/query/plan/time_order_aggregate.h"
 
 namespace stdb {
 namespace qp {
-
-using namespace storage;
-
-/**
- * Tier-1 operator
- */
-struct ProcessingPrelude {
-  virtual ~ProcessingPrelude() = default;
-  //! Compute processing step result (list of low level operators)
-  virtual common::Status apply(const ColumnStore& cstore) = 0;
-  //! Get result of the processing step
-  virtual common::Status extract_result(std::vector<std::unique_ptr<RealValuedOperator>>* dest) = 0;
-  //! Get result of the processing step
-  virtual common::Status extract_result(std::vector<std::unique_ptr<AggregateOperator>>* dest) = 0;
-  //! Get result of the processing step
-  virtual common::Status extract_result(std::vector<std::unique_ptr<BinaryDataOperator>>* dest) = 0;
-  //! Get debug info of the processing step
-  virtual boost::property_tree::ptree debug_info() const = 0;
-};
-
-/**
- * Tier-N operator (materializer)
- */
-struct MaterializationStep {
-  virtual ~MaterializationStep() = default;
-
-  //! Compute processing step result (list of low level operators)
-  virtual common::Status apply(ProcessingPrelude* prelude) = 0;
-
-  /**
-   * Get result of the processing step, this method should add cardinality() elements
-   * to the `dest` array.
-   */
-  virtual common::Status extract_result(std::unique_ptr<ColumnMaterializer>* dest) = 0;
-
-  /**
-   * Get the debug info of MaterializationStep
-   */
-  virtual boost::property_tree::ptree debug_info() const = 0;
-};
-
-struct ScanProcessingStep : ProcessingPrelude {
-  std::vector<std::unique_ptr<RealValuedOperator>> scanlist_;
-  Timestamp begin_;
-  Timestamp end_;
-  std::vector<ParamId> ids_;
-
-  template<class T>
-  ScanProcessingStep(Timestamp begin, Timestamp end, T&& t)
-    : begin_(begin)
-      , end_(end)
-      , ids_(std::forward<T>(t))
-  {
-  }
-
-  boost::property_tree::ptree debug_info() const override {
-    boost::property_tree::ptree tree;
-    tree.add("name", "ScanProcessingStep");
-    return tree;
-  }
-
-  virtual common::Status apply(const ColumnStore& cstore) {
-    return cstore.scan(ids_, begin_, end_, &scanlist_);
-  }
-
-  virtual common::Status extract_result(std::vector<std::unique_ptr<RealValuedOperator>>* dest) {
-    if (scanlist_.empty()) {
-      return common::Status::NoData();
-    }
-    *dest = std::move(scanlist_);
-    return common::Status::Ok();
-  }
-
-  virtual common::Status extract_result(std::vector<std::unique_ptr<AggregateOperator>>* dest) {
-    return common::Status::NoData();
-  }
-
-  virtual common::Status extract_result(std::vector<std::unique_ptr<BinaryDataOperator>>* dest) {
-    return common::Status::NoData();
-  }
-};
-
-struct ScanEventsProcessingStep : ProcessingPrelude {
-  std::vector<std::unique_ptr<BinaryDataOperator>> scanlist_;
-  Timestamp begin_;
-  Timestamp end_;
-  std::vector<ParamId> ids_;
-  std::string regex_;
-
-  //! C-tor (1), create scan without filter
-  template<class T>
-  ScanEventsProcessingStep(Timestamp begin, Timestamp end, T&& t)
-        : begin_(begin)
-        , end_(end)
-        , ids_(std::forward<T>(t))
-  {
-  }
-
-  //! C-tor (2), create scan with filter
-  template<class T>
-  ScanEventsProcessingStep(Timestamp begin, Timestamp end, const std::string& exp, T&& t)
-        : begin_(begin)
-        , end_(end)
-        , ids_(std::forward<T>(t))
-        , regex_(exp)
-  {
-  }
-
-  boost::property_tree::ptree debug_info() const override {
-    boost::property_tree::ptree tree;
-    tree.add("name", "ScanEventsProcessingStep");
-    return tree;
-  }
-
-  virtual common::Status apply(const ColumnStore& cstore) {
-    if (!regex_.empty()) {
-      return cstore.filter_events(ids_, begin_, end_, regex_, &scanlist_);
-    }
-    return cstore.scan_events(ids_, begin_, end_, &scanlist_);
-  }
-
-  virtual common::Status extract_result(std::vector<std::unique_ptr<BinaryDataOperator>>* dest) {
-    if (scanlist_.empty()) {
-      return common::Status::NoData();
-    }
-    *dest = std::move(scanlist_);
-    return common::Status::Ok();
-  }
-
-  virtual common::Status extract_result(std::vector<std::unique_ptr<RealValuedOperator>>* dest) {
-    return common::Status::NoData();
-  }
-
-  virtual common::Status extract_result(std::vector<std::unique_ptr<AggregateOperator>>* dest) {
-    return common::Status::NoData();
-  }
-};
-
-struct FilterProcessingStep : ProcessingPrelude {
-  std::vector<std::unique_ptr<RealValuedOperator>> scanlist_;
-  Timestamp begin_;
-  Timestamp end_;
-  std::map<ParamId, ValueFilter> filters_;
-  std::vector<ParamId> ids_;
-
-  template<class T>
-  FilterProcessingStep(Timestamp begin,
-                       Timestamp end,
-                       const std::vector<ValueFilter>& flt,
-                       T&& t)
-        : begin_(begin)
-        , end_(end)
-        , filters_()
-        , ids_(std::forward<T>(t))
-  {
-    for (size_t ix = 0; ix < ids_.size(); ix++) {
-      ParamId id = ids_[ix];
-      const ValueFilter& filter = flt[ix];
-      filters_.insert(std::make_pair(id, filter));
-    }
-  }
-
-  boost::property_tree::ptree debug_info() const override {
-    boost::property_tree::ptree tree;
-    tree.add("name", "FilterProcessingStep");
-    tree.add("begin", begin_);
-    tree.add("end", end_);
-
-    boost::property_tree::ptree array;
-    for (auto id : ids_) {
-      boost::property_tree::ptree elem;
-      elem.add("id", id);
-      auto iter = filters_.find(id);
-      elem.add("filter", iter->second.debug_string());
-      array.push_back(boost::property_tree::ptree::value_type("", elem));
-    }
-    tree.add_child("ids", array);
-    return tree;
-  }
-
-  virtual common::Status apply(const ColumnStore& cstore) {
-    return cstore.filter(ids_, begin_, end_, filters_, &scanlist_);
-  }
-
-  virtual common::Status extract_result(std::vector<std::unique_ptr<RealValuedOperator>>* dest) {
-    if (scanlist_.empty()) {
-      return common::Status::NoData();
-    }
-    *dest = std::move(scanlist_);
-    return common::Status::Ok();
-  }
-
-  virtual common::Status extract_result(std::vector<std::unique_ptr<AggregateOperator>>* dest) {
-    return common::Status::NoData();
-  }
-
-  virtual common::Status extract_result(std::vector<std::unique_ptr<BinaryDataOperator>>* dest) {
-    return common::Status::NoData();
-  }
-};
-
-struct AggregateProcessingStep : ProcessingPrelude {
-  std::vector<std::unique_ptr<AggregateOperator>> agglist_;
-  Timestamp begin_;
-  Timestamp end_;
-  std::vector<ParamId> ids_;
-
-  template<class T>
-  AggregateProcessingStep(Timestamp begin, Timestamp end, T&& t)
-        : begin_(begin)
-        , end_(end)
-        , ids_(std::forward<T>(t))
-  {
-  }
-
-  boost::property_tree::ptree debug_info() const override {
-    boost::property_tree::ptree tree;
-    tree.add("name", "AggregateProcessingStep");
-    return tree;
-  }
-
-  virtual common::Status apply(const ColumnStore& cstore) {
-    return cstore.aggregate(ids_, begin_, end_, &agglist_);
-  }
-
-  virtual common::Status extract_result(std::vector<std::unique_ptr<RealValuedOperator>>* dest) {
-    return common::Status::NoData();
-  }
-
-  virtual common::Status extract_result(std::vector<std::unique_ptr<AggregateOperator>>* dest) {
-    if (agglist_.empty()) {
-      return common::Status::NoData();
-    }
-    *dest = std::move(agglist_);
-    return common::Status::Ok();
-  }
-
-  virtual common::Status extract_result(std::vector<std::unique_ptr<BinaryDataOperator>>* dest) {
-    return common::Status::NoData();
-  }
-};
-
-
-// Convert AggregateOperator into RealValuedOperator
-class GroupAggregateConverter : public RealValuedOperator {
-  std::unique_ptr<AggregateOperator> op_;
-  AggregationFunction func_;
- 
- public:
-  GroupAggregateConverter(AggregationFunction func, std::unique_ptr<AggregateOperator> op)
-      : op_(std::move(op)), func_(func) { }
-
-  std::tuple<common::Status, size_t> read(Timestamp *destts, double *destval, size_t size) override {
-    size_t pos = 0;
-    while (pos < size) {
-      common::Status status;
-      size_t ressz;
-      Timestamp ts;
-      AggregationResult xs;
-      std::tie(status, ressz) = op_->read(&ts, &xs, 1);
-      if (ressz == 1) {
-        destts[pos] = ts;
-        destval[pos] = TupleOutputUtils::get(xs, func_);
-        pos++;
-      } else if (status.IsOk() || status.Code() == common::Status::kNoData) {
-        return std::make_tuple(status, pos);
-      } else {
-        return std::make_tuple(status, 0);
-      }
-    }
-    return std::make_tuple(common::Status::Ok(), pos);
-  }
-
-  Direction get_direction() override {
-    return AggregateOperator::Direction::FORWARD == op_->get_direction()
-        ? RealValuedOperator::Direction::FORWARD
-        : RealValuedOperator::Direction::BACKWARD;
-  }
-};
-
-
-struct GroupAggregateProcessingStep : ProcessingPrelude {
-  std::vector<std::unique_ptr<AggregateOperator>> agglist_;
-  Timestamp begin_;
-  Timestamp end_;
-  Timestamp step_;
-  std::vector<ParamId> ids_;
-  AggregationFunction fn_;
-
-  template<class T>
-  GroupAggregateProcessingStep(Timestamp begin, Timestamp end, Timestamp step, T&& t, AggregationFunction fn=AggregationFunction::FIRST)
-        : begin_(begin)
-        , end_(end)
-        , step_(step)
-        , ids_(std::forward<T>(t))
-        , fn_(fn) { }
-
-  boost::property_tree::ptree debug_info() const override {
-    boost::property_tree::ptree tree;
-    tree.add("name", "GroupAggregateProcessingStep");
-    return tree;
-  }
-
-  virtual common::Status apply(const ColumnStore& cstore) {
-    return cstore.group_aggregate(ids_, begin_, end_, step_, &agglist_);
-  }
-
-  virtual common::Status extract_result(std::vector<std::unique_ptr<RealValuedOperator>>* dest) {
-    if (agglist_.empty()) {
-      return common::Status::NoData();
-    }
-    dest->clear();
-    for (auto&& it: agglist_) {
-      std::unique_ptr<RealValuedOperator> op;
-      op.reset(new GroupAggregateConverter(fn_, std::move(it)));
-      dest->push_back(std::move(op));
-    }
-    agglist_.clear();
-    return common::Status::Ok();
-  }
-
-  virtual common::Status extract_result(std::vector<std::unique_ptr<AggregateOperator>>* dest) {
-    if (agglist_.empty()) {
-      return common::Status::NoData();
-    }
-    *dest = std::move(agglist_);
-    return common::Status::Ok();
-  }
-
-  virtual common::Status extract_result(std::vector<std::unique_ptr<BinaryDataOperator>>* dest) {
-    return common::Status::NoData();
-  }
-};
-
-struct GroupAggregateFilterProcessingStep : ProcessingPrelude {
-  std::vector<std::unique_ptr<AggregateOperator>> agglist_;
-  Timestamp begin_;
-  Timestamp end_;
-  Timestamp step_;
-  std::vector<ParamId> ids_;
-  std::map<ParamId, AggregateFilter> filters_;
-  AggregationFunction fn_;
-
-  template<class T>
-  GroupAggregateFilterProcessingStep(Timestamp begin,
-                                     Timestamp end,
-                                     Timestamp step,
-                                     const std::vector<AggregateFilter>& flt,
-                                     T&& t,
-                                     AggregationFunction fn=AggregationFunction::FIRST)
-        : begin_(begin)
-        , end_(end)
-        , step_(step)
-        , ids_(std::forward<T>(t))
-        , fn_(fn)
-  {
-    for (size_t ix = 0; ix < ids_.size(); ix++) {
-      ParamId id = ids_[ix];
-      const auto& filter = flt[ix];
-      filters_.insert(std::make_pair(id, filter));
-    }
-  }
-
-  boost::property_tree::ptree debug_info() const override {
-    boost::property_tree::ptree tree;
-    tree.add("name", "GroupAggregateFilterProcessingStep");
-    return tree;
-  }
-
-  virtual common::Status apply(const ColumnStore& cstore) {
-    return cstore.group_aggfilter(ids_, begin_, end_, step_, filters_, &agglist_);
-  }
-
-  virtual common::Status extract_result(std::vector<std::unique_ptr<RealValuedOperator>>* dest) {
-    if (agglist_.empty()) {
-      return common::Status::NoData();
-    }
-    dest->clear();
-    for (auto&& it: agglist_) {
-      std::unique_ptr<RealValuedOperator> op;
-      op.reset(new GroupAggregateConverter(fn_, std::move(it)));
-      dest->push_back(std::move(op));
-    }
-    agglist_.clear();
-    return common::Status::Ok();
-  }
-
-  virtual common::Status extract_result(std::vector<std::unique_ptr<AggregateOperator>>* dest) {
-    if (agglist_.empty()) {
-      return common::Status::NoData();
-    }
-    *dest = std::move(agglist_);
-    return common::Status::Ok();
-  }
-
-  virtual common::Status extract_result(std::vector<std::unique_ptr<BinaryDataOperator>>* dest) {
-    return common::Status::NoData();
-  }
-};
-
-// -------------------------------- //
-//              Tier-2              //
-// -------------------------------- //
-namespace detail {
-
-template <OrderBy order, class OperatorT>
-struct MergeMaterializerTraits;
-
-template <>
-struct MergeMaterializerTraits<OrderBy::SERIES, RealValuedOperator> {
-  typedef MergeMaterializer<SeriesOrder> Materializer;
-};
-
-template <>
-struct MergeMaterializerTraits<OrderBy::TIME, RealValuedOperator> {
-  typedef MergeMaterializer<TimeOrder> Materializer;
-};
-
-template <>
-struct MergeMaterializerTraits<OrderBy::SERIES, BinaryDataOperator> {
-  typedef MergeEventMaterializer<EventSeriesOrder> Materializer;
-};
-
-template <>
-struct MergeMaterializerTraits<OrderBy::TIME, BinaryDataOperator> {
-  typedef MergeEventMaterializer<EventTimeOrder> Materializer;
-};
-
-}  // namespace detail
-
-/**
- * Merge several series (order by series).
- * Used in scan query.
- */
-template <OrderBy order, class OperatorT = RealValuedOperator>
-struct MergeBy : MaterializationStep {
-  std::vector<ParamId> ids_;
-  std::unique_ptr<ColumnMaterializer> mat_;
-
-  template<class IdVec>
-  MergeBy(IdVec&& ids) : ids_(std::forward<IdVec>(ids)) { }
-
-  boost::property_tree::ptree debug_info() const override {
-    boost::property_tree::ptree tree;
-    tree.add("name", "MergeBy");
-    boost::property_tree::ptree array;
-    for (auto id : ids_) {
-      boost::property_tree::ptree elem;
-      elem.add("id", id);
-      array.push_back(boost::property_tree::ptree::value_type("", elem));
-    }
-    tree.add_child("ids", array);
-    switch (order) {
-      case OrderBy::SERIES: {
-        tree.add("order", "SERIES");
-      } break;
-      case OrderBy::TIME: {
-        tree.add("orderby", "TIME");
-      } break;
-    }
-    return tree;
-  }
-
-  common::Status apply(ProcessingPrelude* prelude) {
-    std::vector<std::unique_ptr<OperatorT>> iters;
-    auto status = prelude->extract_result(&iters);
-    if (!status.IsOk()) {
-      return status;
-    }
-    typedef typename detail::MergeMaterializerTraits<order, OperatorT>::Materializer Merger;
-    mat_.reset(new Merger(std::move(ids_), std::move(iters)));
-    return common::Status::Ok();
-  }
-
-  common::Status extract_result(std::unique_ptr<ColumnMaterializer> *dest) {
-    if (!mat_) {
-      return common::Status::NoData();
-    }
-    *dest = std::move(mat_);
-    return common::Status::Ok();
-  }
-};
-
-namespace detail {
-
-template <class OperatorT>
-struct ChainMaterializerTraits;
-
-template <>
-struct ChainMaterializerTraits<RealValuedOperator> {
-  typedef ChainMaterializer Materializer;
-};
-
-template <>
-struct ChainMaterializerTraits<BinaryDataOperator> {
-  typedef EventChainMaterializer Materializer;
-};
-
-}  // namespace detail
-
-template <class OperatorT = RealValuedOperator>
-struct Chain : MaterializationStep {
-  std::vector<ParamId> ids_;
-  std::unique_ptr<ColumnMaterializer> mat_;
-
-  template<class IdVec>
-  Chain(IdVec&& vec) : ids_(std::forward<IdVec>(vec)) { }
-
-  boost::property_tree::ptree debug_info() const override {
-    boost::property_tree::ptree tree;
-    tree.add("name", "Chain");
-    return tree;
-  }
-
-  common::Status apply(ProcessingPrelude *prelude) {
-    std::vector<std::unique_ptr<OperatorT>> iters;
-    auto status = prelude->extract_result(&iters);
-    if (!status.IsOk()) {
-      return status;
-    }
-    typedef typename detail::ChainMaterializerTraits<OperatorT>::Materializer Materializer;
-    mat_.reset(new Materializer(std::move(ids_), std::move(iters)));
-    return common::Status::Ok();
-  }
-
-  common::Status extract_result(std::unique_ptr<ColumnMaterializer> *dest) {
-    if (!mat_) {
-      return common::Status::NoData();
-    }
-    *dest = std::move(mat_);
-    return common::Status::Ok();
-  }
-};
-
-/**
- * Aggregate materializer.
- * Accepts the list of ids and the list of aggregate operators.
- * Maps each id to the corresponding operators 1-1.
- * All ids should be different.
- */
-struct Aggregate : MaterializationStep {
-  std::vector<ParamId> ids_;
-  std::vector<AggregationFunction> fn_;
-  std::unique_ptr<ColumnMaterializer> mat_;
-
-  template<class IdVec, class FuncVec>
-  Aggregate(IdVec&& vec, FuncVec&& fn)
-     : ids_(std::forward<IdVec>(vec))
-      , fn_(std::forward<FuncVec>(fn)) { }
-
-  boost::property_tree::ptree debug_info() const override {
-    boost::property_tree::ptree tree;
-    tree.add("name", "Aggregate");
-    return tree;
-  }
-
-  common::Status apply(ProcessingPrelude *prelude) {
-    std::vector<std::unique_ptr<AggregateOperator>> iters;
-    auto status = prelude->extract_result(&iters);
-    if (!status.IsOk()) {
-      return status;
-    }
-    mat_.reset(new AggregateMaterializer(std::move(ids_), std::move(iters), std::move(fn_)));
-    return common::Status::Ok();
-  }
-
-  common::Status extract_result(std::unique_ptr<ColumnMaterializer> *dest) {
-    if (!mat_) {
-      return common::Status::NoData();
-    }
-    *dest = std::move(mat_);
-    return common::Status::Ok();
-  }
-};
-
-/**
- * Combines the aggregate operators.
- * Accepts list of ids (shouldn't be different) and list of aggregate
- * operators. Maps each id to operator and then combines operators
- * with the same id (used to implement aggregate + group-by).
- */
-struct AggregateCombiner : MaterializationStep {
-  std::vector<ParamId> ids_;
-  std::vector<AggregationFunction> fn_;
-  std::unique_ptr<ColumnMaterializer> mat_;
-
-  template <class IdVec, class FuncVec>
-  AggregateCombiner(IdVec&& vec, FuncVec&& fn)
-      : ids_(std::forward<IdVec>(vec))
-        , fn_(std::forward<FuncVec>(fn)) { }
-
-  boost::property_tree::ptree debug_info() const override {
-    boost::property_tree::ptree tree;
-    tree.add("name", "AggregateCombiner");
-    return tree;
-  }
-
-  common::Status apply(ProcessingPrelude *prelude) {
-    std::vector<std::unique_ptr<AggregateOperator>> iters;
-    auto status = prelude->extract_result(&iters);
-    if (status != common::Status::Ok()) {
-      return status;
-    }
-    std::vector<std::unique_ptr<AggregateOperator>> agglist;
-    std::map<ParamId, std::vector<std::unique_ptr<AggregateOperator>>> groupings;
-    std::map<ParamId, AggregationFunction> functions;
-    for (size_t i = 0; i < ids_.size(); i++) {
-      auto id = ids_.at(i);
-      auto it = std::move(iters.at(i));
-      groupings[id].push_back(std::move(it));
-      functions[id] = fn_.at(i);
-    }
-    std::vector<ParamId> ids;
-    std::vector<AggregationFunction> fns;
-    for (auto& kv: groupings) {
-      auto& vec = kv.second;
-      ids.push_back(kv.first);
-      std::unique_ptr<CombineAggregateOperator> it(new CombineAggregateOperator(std::move(vec)));
-      agglist.push_back(std::move(it));
-      fns.push_back(functions[kv.first]);
-    }
-    mat_.reset(new AggregateMaterializer(std::move(ids),
-                                         std::move(agglist),
-                                         std::move(fns)));
-    return common::Status::Ok();
-  }
-
-  common::Status extract_result(std::unique_ptr<ColumnMaterializer> *dest) {
-    if (!mat_) {
-      return common::Status::NoData();
-    }
-    *dest = std::move(mat_);
-    return common::Status::Ok();
-  }
-};
-
-
-template <OrderBy order>
-struct GroupAggregateCombiner_Initializer;
-
-template <>
-struct GroupAggregateCombiner_Initializer<OrderBy::SERIES> {
-  static std::unique_ptr<ColumnMaterializer> make_materializer(
-      std::vector<ParamId>&& ids,
-      std::vector<std::unique_ptr<AggregateOperator>>&& agglist,
-      const std::vector<AggregationFunction>& fn)
-  {
-    std::unique_ptr<ColumnMaterializer> mat;
-    mat.reset(new SeriesOrderAggregateMaterializer(std::move(ids), std::move(agglist), fn));
-    return mat;
-  }
-};
-
-template <>
-struct GroupAggregateCombiner_Initializer<OrderBy::TIME> {
-  static std::unique_ptr<ColumnMaterializer> make_materializer(
-      std::vector<ParamId>&& ids,
-      std::vector<std::unique_ptr<AggregateOperator>>&& agglist,
-      const std::vector<AggregationFunction>& fn)
-  {
-    std::vector<ParamId> tmpids(ids);
-    std::vector<std::unique_ptr<AggregateOperator>> tmpiters(std::move(agglist));
-    std::unique_ptr<ColumnMaterializer> mat;
-    mat.reset(new TimeOrderAggregateMaterializer(tmpids, tmpiters, fn));
-    return mat;
-  }
-};
-
-/**
- * Combines the group-aggregate operators.
- * Accepts list of ids (shouldn't be different) and list of aggregate
- * operators. Maps each id to operator and then combines operators
- * with the same id (to implement group-aggregate + group/pivot-by-tag).
- */
-template <OrderBy order>
-struct GroupAggregateCombiner : MaterializationStep {
-  std::vector<ParamId> ids_;
-  std::vector<AggregationFunction> fn_;
-  std::unique_ptr<ColumnMaterializer> mat_;
-
-  template<class IdVec, class FuncVec>
-  GroupAggregateCombiner(IdVec&& vec, FuncVec&& fn)
-        : ids_(std::forward<IdVec>(vec))
-        , fn_(std::forward<FuncVec>(fn)) { }
-
-  boost::property_tree::ptree debug_info() const override {
-    boost::property_tree::ptree tree;
-    tree.add("name", "GroupAggregateCombiner");
-    return tree;
-  }
-
-  common::Status apply(ProcessingPrelude *prelude) {
-    std::vector<std::unique_ptr<AggregateOperator>> iters;
-    auto status = prelude->extract_result(&iters);
-    if (status != common::Status::Ok()) {
-      return status;
-    }
-    std::vector<std::unique_ptr<AggregateOperator>> agglist;
-    std::map<ParamId, std::vector<std::unique_ptr<AggregateOperator>>> groupings;
-    for (size_t i = 0; i < ids_.size(); i++) {
-      auto id = ids_.at(i);
-      auto it = std::move(iters.at(i));
-      groupings[id].push_back(std::move(it));
-    }
-    std::vector<ParamId> ids;
-    for (auto& kv: groupings) {
-      auto& vec = kv.second;
-      ids.push_back(kv.first);
-      std::unique_ptr<FanInAggregateOperator> it(new FanInAggregateOperator(std::move(vec)));
-      agglist.push_back(std::move(it));
-    }
-    mat_ = GroupAggregateCombiner_Initializer<order>::make_materializer(std::move(ids),
-                                                                        std::move(agglist),
-                                                                        fn_);
-    return common::Status::Ok();
-  }
-
-  common::Status extract_result(std::unique_ptr<ColumnMaterializer> *dest) {
-    if (!mat_) {
-      return common::Status::NoData();
-    }
-    *dest = std::move(mat_);
-    return common::Status::Ok();
-  }
-};
-
-/**
- * Joins several operators into one.
- * Number of joined operators is defined by the cardinality.
- * Number of ids should be `cardinality` times smaller than number
- * of operators because every `cardinality` operators are joined into
- * one.
- */
-struct Join : MaterializationStep {
-  std::vector<ParamId> ids_;
-  int cardinality_;
-  OrderBy order_;
-  Timestamp begin_;
-  Timestamp end_;
-  std::unique_ptr<ColumnMaterializer> mat_;
-
-  template<class IdVec>
-  Join(IdVec&& vec, int cardinality, OrderBy order, Timestamp begin, Timestamp end)
-        : ids_(std::forward<IdVec>(vec))
-        , cardinality_(cardinality)
-        , order_(order)
-        , begin_(begin)
-        , end_(end) { }
-
-  boost::property_tree::ptree debug_info() const override {
-    boost::property_tree::ptree tree;
-    tree.add("name", "Join");
-    return tree;
-  }
-
-  common::Status apply(ProcessingPrelude *prelude) {
-    int inc = cardinality_;
-    std::vector<std::unique_ptr<RealValuedOperator>> scanlist;
-    auto status = prelude->extract_result(&scanlist);
-    if (status != common::Status::Ok()) {
-      return status;
-    }
-    std::vector<std::unique_ptr<ColumnMaterializer>> iters;
-    for (size_t i = 0; i < ids_.size(); i++) {
-      // ids_ contain ids of the joined series that corresponds
-      // to the names in the series matcher
-      std::vector<std::unique_ptr<RealValuedOperator>> joined;
-      std::vector<ParamId> ids;
-      for (int j = 0; j < inc; j++) {
-        // `inc` number of storage level operators correspond to one
-        // materializer
-        auto ix = i*inc + j;
-        joined.push_back(std::move(scanlist.at(ix)));
-        ids.push_back(ix);
-      }
-      std::unique_ptr<ColumnMaterializer> it;
-      it.reset(new JoinMaterializer(std::move(ids), std::move(joined), ids_.at(i)));
-      iters.push_back(std::move(it));
-    }
-    if (order_ == OrderBy::SERIES) {
-      mat_.reset(new JoinConcatMaterializer(std::move(iters)));
-    } else {
-      bool forward = begin_ < end_;
-      typedef MergeJoinMaterializer<MergeJoinUtil::OrderByTimestamp> Materializer;
-      mat_.reset(new Materializer(std::move(iters), forward));
-    }
-    return common::Status::Ok();
-  }
-
-  common::Status extract_result(std::unique_ptr<ColumnMaterializer> *dest) {
-    if (!mat_) {
-      return common::Status::NoData();
-    }
-    *dest = std::move(mat_);
-    return common::Status::Ok();
-  }
-};
-
-/**
- * Merges several group-aggregate operators by chaining
- */
-struct SeriesOrderAggregate : MaterializationStep {
-  std::vector<ParamId> ids_;
-  std::vector<AggregationFunction> fn_;
-  std::unique_ptr<ColumnMaterializer> mat_;
-
-  template <class IdVec, class FnVec>
-  SeriesOrderAggregate(IdVec&& vec, FnVec&& fn)
-        : ids_(std::forward<IdVec>(vec))
-        , fn_(std::forward<FnVec>(fn)) { }
-
-  boost::property_tree::ptree debug_info() const override {
-    boost::property_tree::ptree tree;
-    tree.add("name", "SeriesOrderAggregate");
-    return tree;
-  }
-
-  common::Status apply(ProcessingPrelude *prelude) {
-    std::vector<std::unique_ptr<AggregateOperator>> iters;
-    auto status = prelude->extract_result(&iters);
-    if (status != common::Status::Ok()) {
-      return status;
-    }
-    mat_.reset(new SeriesOrderAggregateMaterializer(std::move(ids_), std::move(iters), fn_));
-    return common::Status::Ok();
-  }
-
-  common::Status extract_result(std::unique_ptr<ColumnMaterializer> *dest) {
-    if (!mat_) {
-      return common::Status::NoData();
-    }
-    *dest = std::move(mat_);
-    return common::Status::Ok();
-  }
-};
-
-struct TimeOrderAggregate : MaterializationStep {
-  std::vector<ParamId> ids_;
-  std::vector<AggregationFunction> fn_;
-  std::unique_ptr<ColumnMaterializer> mat_;
-
-  template<class IdVec, class FnVec>
-  TimeOrderAggregate(IdVec&& vec, FnVec&& fn)
-        : ids_(std::forward<IdVec>(vec))
-        , fn_(std::forward<FnVec>(fn)) { }
-
-  boost::property_tree::ptree debug_info() const {
-    boost::property_tree::ptree tree;
-    tree.add("name", "TimeOrderAggregate");
-    return tree;
-  }
-
-  common::Status apply(ProcessingPrelude *prelude) {
-    std::vector<std::unique_ptr<AggregateOperator>> iters;
-    auto status = prelude->extract_result(&iters);
-    if (status != common::Status::Ok()) {
-      return status;
-    }
-    mat_.reset(new TimeOrderAggregateMaterializer(ids_, iters, fn_));
-    return common::Status::Ok();
-  }
-
-  common::Status extract_result(std::unique_ptr<ColumnMaterializer> *dest) {
-    if (!mat_) {
-      return common::Status::NoData();
-    }
-    *dest = std::move(mat_);
-    return common::Status::Ok();
-  }
-};
 
 struct TwoStepQueryPlan : IQueryPlan {
   std::unique_ptr<ProcessingPrelude> prelude_;
